Refactor modal wrapper  using composition
Issue: ModalWrapper component had too many props that was only related to child components
Solution: Apply composition to ModalWrapper
Risk: Further refactoring required to make app work
Reviewed by: Jason Ko
