+++ frontend/src/components/common/modalWrapper/modalWrapper.jsx
@@ -1,16 +1,7 @@
 import React from "react";
-import HabitDetailModal from "../../habitDetailModal/habitDetailModal";
-import IdentityDetailModal from "../../identityDetailModal/identityDetailModal";
-import NewHabitModal from "../../newHabitModal/newHabitModal";
-import NewIdentityModal from "../../newIdentityModal/newIdentityModal";
 import styles from "./modalWrapper.module.css";
 
-const ModalWrapper = ({
-  modalType,
-  closeModal,
-  addNewHabit,
-  addNewIdentity,
-}) => {
+const ModalWrapper = ({ children, closeModal }) => {
   const onBackgroundClickHandler = (e) => {
     if (e.target.id === "modalWrapper") {
       closeModal();
@@ -19,25 +10,11 @@ const ModalWrapper = ({
 
   return (
     <div
-      onClick={onBackgroundClickHandler}
       id="modalWrapper"
       className={styles.container}
+      onClick={onBackgroundClickHandler}
     >
-      {modalType === "habit" && (
-        <NewHabitModal closeModal={closeModal} addNewHabit={addNewHabit} />
-      )}
-      {modalType === "identity" && (
-        <NewIdentityModal
-          closeModal={closeModal}
-          addNewIdentity={addNewIdentity}
-        />
-      )}
-      {modalType === "habit_detail" && (
-        <HabitDetailModal closeModal={closeModal} />
-      )}
-      {modalType === "identity_detail" && (
-        <IdentityDetailModal closeModal={closeModal} />
-      )}
+      {children}
     </div>
   );
 };
